Move `LogSubscriber` to the tokio-trace-log crate
