+++ tokio-trace-log/Cargo.toml
@@ -6,3 +6,6 @@ authors = ["Eliza Weisman <eliza@buoyant.io>"]
 [dependencies]
 tokio-trace = { path = "../tokio-trace" }
 log = "0.4"
+
+[dev-dependencies]
+env_logger = "0.5"
