+++ tokio-trace-log/src/lib.rs
@@ -23,7 +23,7 @@ extern crate tokio_trace;
 extern crate log;
 
 use std::{io, time::Instant};
-use tokio_trace::{Subscriber, Event};
+use tokio_trace::{Subscriber, Event, SpanData, Meta};
 
 /// Format a log record as a trace event in the current span.
 pub fn format_trace(record: &log::Record) -> io::Result<()> {
@@ -40,12 +40,39 @@ pub fn format_trace(record: &log::Record) -> io::Result<()> {
     Ok(())
 }
 
+pub fn meta_to_log_metadata<'a>(meta: &tokio_trace::Meta<'a>) -> log::Metadata<'a> {
+    log::Metadata::builder()
+        .level(meta.level)
+        .target(meta.target.unwrap_or(""))
+        .build()
+}
+
+pub fn log_record_to_meta<'a>(record: &log::Record<'a>) -> Meta<'a> {
+    Meta {
+        name: None,
+        target: Some(record.target()),
+        level: record.level(),
+        module_path: record
+            .module_path()
+            // TODO: make symmetric
+            .unwrap_or_else(|| record.target()),
+        line: record.line().unwrap_or(0),
+        file: record.file().unwrap_or("???"),
+        field_names: &[],
+    }
+}
+
 /// A simple "logger" that converts all log records into `tokio_trace` `Event`s,
 /// with an optional level filter.
 #[derive(Debug)]
 pub struct SimpleTraceLogger {
     filter: log::LevelFilter,
 }
+
+/// A `tokio_trace` subscriber that logs all recorded trace events.
+pub struct LogSubscriber;
+
+// ===== impl SimpleTraceLogger =====
 
 impl SimpleTraceLogger {
     pub fn with_filter(filter: log::LevelFilter) -> Self {
@@ -71,4 +98,53 @@ impl log::Log for SimpleTraceLogger {
     }
 
     fn flush(&self) {}
+}
+
+// ===== impl LogSubscriber =====
+
+impl LogSubscriber {
+    pub fn new() -> Self {
+        LogSubscriber
+    }
+}
+
+impl Subscriber for LogSubscriber {
+    fn enabled(&self, metadata: &Meta) -> bool {
+        log::logger().enabled(&meta_to_log_metadata(metadata))
+    }
+
+    fn observe_event<'event, 'meta: 'event>(&self, event: &'event Event<'event, 'meta>) {
+        let fields = event.debug_fields();
+        let meta = meta_to_log_metadata(event.meta);
+        let logger = log::logger();
+        let parents = event.parents().filter_map(SpanData::name).collect::<Vec<_>>();
+        if logger.enabled(&meta) {
+            logger.log(
+                &log::Record::builder()
+                    .metadata(meta)
+                    .module_path(Some(event.meta.module_path))
+                    .file(Some(event.meta.file))
+                    .line(Some(event.meta.line))
+                    .args(format_args!(
+                        "[{}] {:?} {}",
+                        parents.join(":"),
+                        fields,
+                        event.message
+                    )).build(),
+            );
+        }
+    }
+
+    fn enter(&self, span: &SpanData, _at: Instant) {
+        let logger = log::logger();
+        logger.log(&log::Record::builder()
+            .args(format_args!("-> {:?}", span.name()))
+            .build()
+        )
+    }
+
+    fn exit(&self, span: &SpanData, _at: Instant) {
+        let logger = log::logger();
+        logger.log(&log::Record::builder().args(format_args!("<- {:?}", span.name())).build())
+    }
 }
