+++ tokio-trace/Cargo.toml
@@ -8,6 +8,7 @@ futures = "0.1"
 log = "0.4"
 
 [dev-dependencies]
-env_logger = "0.5"
 ansi_term = "0.11"
 humantime = "1.1.1"
+env_logger = "0.5"
+tokio-trace-log = { path = "../tokio-trace-log" }
